Update LDAP Filter
Updated the LDAP filter to only return objects that have a `thumbnailphoto` property with any value. I was recently in an environment where the built-in search base returned multiple objects. The for loop over the `mySearcher.FindAll()` results errors out and doesn't recover because the first returned object did not have a `thumbnailphoto` property.